API-12090 adjust differences between v3.3 and v3.4 for agent
